Gitaly-ruby server should not depend on the global state
Configuration of the gitaly-ruby server depends on the
global config.Config variable. As we are on the road to
removing it this change break this dependency by introducing
a constructor function that accepts required configuration.
That change requires refactoring of the gitaly-ruby test
instance setup - now it should be created only after
configuration is properly set.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699