make AddAddress synchronous
AddAddress adds an address of a gitaly-ruby worker to the resolver
where gRPC ClientConns get the addresses of available gitaly-ruby
workers. Currently, this happens asynchronously. The operation is
fast and there doesn't seem to be a clear reason why this should be
asynchronous. This commit makes AddAddress wait until the address
has actually been added to the pool of addresses. This helps making
the code deterministic to fix a flaky test.