fix flaky test 'TestNewPostgresListener/invalid_connection'
The test asserts that there's a log statement printed out with the
error that caused the connection failure to Postgres. The error is
sent over a channel to the constructor so an error can be returned
if the first connection attempt fails, likely indicating an invalid
address. The error is sent on the channel prior to logging out the
error though, which causes a race where the constructor might return
and execute the test assertions before the log statement is written.