fix flaky test TestDial_Tracing/stream
This commit fixes TestDial_Tracing/stream which is flaking and failing
the pipeline every now and then.

The ordering of the items is indeterministic due to a race in finishing
two spans. The main test goroutine does not wait for the stream to close
but only receives a single message. Once that message is received, both
the main test goroutine and the stream handler race to finish their spans.
This commit fixes the situation by making the main test goroutine wait for
the stream to close and only then finishing its span.