proxy: Pass down destination to proxying code
The proxying code where the "real magic" happens currently only gets the
destination's peeked header message as well as the stream it's supposed
to proxy on. We'll need more information than this though in a later
commit, so let's refactor the code to instead pass down the complete
Destination structure.

No change in behaviour is expected from this commit.