proxy: Allow setup of error handlers
Originally, the GRPC proxy code only handled a single destination. With
the extension to also handle proxying to secondary nodes, its error
handling started to become kind of weird: we only get a subset of errors
and have no control over how these errors are handled. This brings along
multiple issues: first, we cannot know downstream in the coordinator
whether a node has failed or not and thus cannot know when to schedule
replication jobs. And second, when proxying to any secondary fails while
proxying to the primary succeeds, the proxy code will regard it as an
overall failure.

It's unfortunately hard to pin down exact semantics for error handling
here. One may wish to ignore errors from secondaries, but potentially
not in all cases. E.g. for transactional RPCs, we may only want to
ignore errors from secondaries until we know that its impossible to
reach quorum anymore because too many nodes have failed already. Coding
this behaviour into the proxy handler doesn't really make a lot of
sense.

This commit instead implements a new parameter for proxy destinations,
which is an optional error handler. If set, this function will be called
for all non-EOF errors. This allows both for custom handling of errors
in the coordinator as well as for filtering of errors.

The error handler is not yet used by any callsite.