remote: Convert users and remove package
Now that our localrepo interface provides all functionality of the
remote package, let's convert callsites to use the localrepo interface
instead and remove the now-unused remote package.

While at it, this commit also moves over a testcase for deletion of
local branches to ensure that test coverage doesn't decrease because of
the package deletion.