Makefile: Do not re-clone repos when rebuilding git/libgit2
When the Makefile changes, we always rebuild both git and libgit2
because we must assume that their requested version has changed. This is
perfectly fine, but whenever we do we also throw away their repositories
and just re-clone them anew, which is a complete waste of resources.

Fix this shortcoming by only conditionally cloning the repos if they do
not exist yet. If they do exist, we must make sure to update it
accordingly in case a revision was requested which isn't yet known.

This commit also partly fixes a shortcoming of the current approach: it
isn't possible to directly specify an object ID either as git or libgit2
version as we used `git clone --branch ${GIT_VERSION}` to perform the
clone. It's still not possible for the initial clone as to the best of
my knowledge, git-clone(1) doesn't support directly cloning a specific
commit. This could be fixed by doing a non-shallow clone, but the
performance hit on CI systems is probably not worth it. It now is
possible to specify an object ID when the repository already exists and
we only update it though.