Makefile: Fix libgit2.a poisoning Go build cache
In order to speed up builds, the Go toolchain uses a build cache which
hosts pre-compiled assets. This build cache works well enough in most
cases, but in combination with CGo it has some shortcomings as it also
caches paths of static archives.

Given that the build cache is shared between all Go projects on the same
disk independent of where they're hosted, it means that if two projects
link against a version of a static archive, they'll both use the cached
version of the path. If project A has compiled this archive first, then
project B will also use that same static archive of project A. This can
be easily observed for Gitaly, too:

    1. Erase the current build cache.

    2. Clone Gitaly in a temporary directory and build it.

    3. Delete the temporary directory.

    4. Clone Gitaly in a different temporary directory and build it.

The fourth step will fail with a linker error because the Go toolchain
tries to re-use the old cached path of libgit2.a, which has been removed
in the third step. This isn't only an annoyance which causes the build
to fail, but can potentially also cause us to link against the wrong
libgit2.a.

The problem cannot really be solved in a way that's a 100% satisfying:
force-rebuilding Git2Go doesn't fix the issue as the cached path is part
of gitaly-git2go's build cache. So in order to accurately fix the
problem, we'd have to force-rebuild that command. This is causing a huge
performance hit though during a normal development session and thus
isn't really acceptable.

Instead, this commit simply overrides the Go cache directory so that
it's per-project in our own build directory. This fixes the poisoning
issue because the libgit2.a path shouldn't really change now. And if it
does, it's easy to be solved by doing a simple `make clean`. It
partitions the cache though and thus adds some overhead when building
Gitaly from many different paths, but this should be an acceptable
tradeoff.

The worst part of this workaround is that for normal development flows,
we'll now use two caches: the global cache when the developer manually
executes `go build/test/install`, and the local cache when the developer
uses `make`. But that's probably the better trade-off compared to a
forced rebuild of everything each time one executes `make`.