reconcile missing repositories to assigned nodes
Reconciler currently schedules jobs to keep repositories on assigned
host storages up to date but doesn't replicate the repository to an
assigned host node that is missing the repository entirely.

This bug is due to incorrectly filtering assignments using
`storage_repositories` table's `storage` column. The assigned storage
has no entry for the repository in the table if it is missing the
repository. This then results in comparing to a `null` `storage` column
which fails to find the node assigned. This commit fixes the query to use
the `healthy_storages.storage` as it is guaranteed to not be null when
performing the check. This fixes the issue and correctly identifies the
storage node as assigned even if it is missing a repository and the
`storage_repositories` row.