remove reconciliation job count metric
Reconciliation job count metric records the source and the target
storage. Delete type jobs won't have a source storage which would
require us to change the existing metric. The metric has not been
added to any of the default dashboards, so let's remove it. The
information can be gathered in better detail from the logs as we
print out all scheduled jobs.