praefect: Consider RPC errors when creating replication jobs
Historically, we didn't have any access to error codes of proxying
destinations if proxying to more than one node at once. This is mostly
because the proxy code just conflated all errors into one, which made
proper error handling next to impossible. As a result, the transactional
finalizer cannot and thus doesn't take into account whether any of the
nodes fails, which may lead to spotty coverage when creating replication
jobs.

Fix this shortcoming by setting up error handlers for all proxying
destinations. Like this, we can easily grab any errors they encounter
and put it into a local map, which we can then forward to the
transaction finalizer. Now there's three cases to consider:

    - The primary raised an error. In that case, we cannot help it and
      need to create replication jobs to all nodes.

    - A secondary raised an error. In that case, we need to consider it
      as outdated and need to create a replication job for that specific
      node.

    - Neither of them failed, in which case the node is considered to be
      up-to-date.

Note that currently we still pass the error through as-is even for the
secondaries. We should instead ignore errors on secondaries, but this is
left for a subsequent commit.