praefect: Always create replication jobs for transactional RPCs
The current implementation of determining updated and outdated
secondaries for a transactional RPC is suboptimal as there are several
conditions where a node is not considered to be outdated. As a result,
we wouldn't schedule replication jobs in these scenarios and potentially
the state would be inconsistent across nodes.

This commit addresses most of these shortcomings:

    - When no subtransactions were created, replication targets which
      didn't take part in the transaction weren't considered outdated.

    - If the primary failed the vote and didn't reach quorum, then no
      replication jobs were created. We cannot judge though as there
      might have been on-disk changes which we must replicate.

    - If we failed to get the transaction state, then we simply errored
      and didn't finalize the transaction at all. While this does
      indicate a real error, we must still finalize the transaction and
      err on the safe side by creating replication jobs.

One shortcoming that isn't yet addressed is if any of the nodes fails
the RPC. This change is left for a subsequent commit so that we do the
best we can with the info we've currently got available.