updateref: Capture stderr to silence logs on failing update
When git commands get executed without stderr being captured, we
automatically generate error log messages. While this is implemented as
a help for debugging errors, in many cases it causes us to generate logs
for errors which we know to handle correctly and which are expected.

One such case is in updateref: the code here is inherently racy as any
two processes which modify references may race with each other. It is
thus totally expected that updateref fails, but given that we execute it
without stderr being set up we will log an error regardless.

Fix this issue by capturing git-update-ref(1)'s stderr and returning it
on `Wait()` if any error occurred.