do not use replicas which are pending deletion as replication source
Praefect's reconciler is going to start scheduling jobs to delete
unassigned repositories. With that in place, we should avoid using
repositories that are scheduled for deletion as replication sources.
This commit filters out up to date replicas from being used as a source
if they have been scheduled for deletion.

This should not generally happen anyway, given we only delete unassigned
replicas if all assigned storages (replication targets) are up to date.
This could happen if a storage is unassigned and rapidly another storage
is assigned. The job would fail and a replication would get scheduled from
another node, so this should not be a big problem but we may as well avoid
it anyway.