praefect: Implement tests for failing RPC error proxying
This commit implements a bunch of tests to assert that we correctly
proxy GRPC errors from primary while ignoring any errors from
secondaries. One of the tests is currently broken and will be fixed in a
subsequent commit.

The tests use the MockManager, which was missing two functions to get
nodes. These have also been mocked such that tests succeed.