praefect: Extract function to determine updated/outdated nodes
The transactional request finalizer is quite hard to follow. This has
led to multiple issues with the current implementation where it doesn't
correctly handle some nodes and won't create replication jobs when
there's an error.

As a preparation for refactorings, this commit pulls out the code to
determine outdated and up-to-date nodes into its own standalone
function and adds a bunch of tests.

No functional change is expected from this commit.