ci: Test against Git v2.30.1
Our merge request pipelines currently only test against Git v2.29.0,
and given that this is our minimum required version, we should continue
testing against it. But in order to ensure that we're compatible against
a broader range of versions and to enable a smooth upgrade path in case
we're going to bump the minimum version, we should also test against
newer git versions.

Add tests against the latest released version to improve our test
coverage.