Removal of config.Config from walker_test.go
Usages of the config.Config replaced with locally
instantiated config struct. MockCounter made public
to be used in the cache_test. The counter is mocked
and as it increments in async way it changes overtime
and expected value memorized could be changed by other
running go routine. That is why mock replaced with a
new instance other go routines don't know about.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699