catfile: Rename `batchCheck` to `batchCheckProcess`
When creating a new `catfile.Batch`, it spawns two processes: `git
cat-file --batch` and `git cat-file --batch-check`. These are used to
read objects and read their info, respectively. But while the struct
hosting logic for the first one is called `batchProcess`, the latter one
is called `batchCheck`. This makes the code kind of confusing to read.

Rename the latter one to `batchCheckProcess` to indicate that it's in
the same spirit as the `batchProcess` and that it hosts a process.