blob: Implement function to find LFS pointers
We're about to port the implementation of the blob service's LFS-related
RPCs to Go. The most important building block for this is to be able to
identify LFS pointers, either by walking a revision graph or by reading
blob IDs directly.

This commit introduces two new functions which implement this with
`findLFSPointersByRevisions()` and `readLFSPointers()`. While the former
one searches the revision graph for LFS pointers, the latter one accepts
a list of potential LFS pointers and returns all which in fact are valid
LFS pointers.

Except for the newly added tests, both functions are not yet used.