blob: Implement slicing of LFS pointers
All three RPCs to find and return LFS pointers are streaming RPCs. Due
to RPC message size limitations, we need to make sure to only send a
subset of found LFS pointers in each message.

This commit introduces a new function `sliceLFSPointers()` which does
exactly that: given a slice of LFS pointers, it'll execute a callback
function with batches of this slice. The maximum slice size is the exact
same as in the Ruby code, namely 100.

Except for the newly added tests, this function is not yet used.