ruby: lfs: Limit blob size when searching for new LFS pointers
Given that LFS is bolted on top of git, git itself doesn't provide any
way to nicely find all LFS pointers. We thus have to resort to
heuristics to identify whether a blob is an LFS pointer or not. The
most important heuristic here is that we limit by blob size: we know
that the maximum size for LFS objects is 200 bytes, so we do not want to
load and scan any larger objects. This is a significant optimization,
given that it allows us to skip loading any biggish objects into memory.

While we do use this optimization when searching all referenced blobs,
we do not when only searching for new pointers in a limited revision
graph. Fix this by using the same filter for both operations, which
should significantly speed getting new pointers.