don't elect replicas pending deletion in per repository elector
Praefect's reconciler is going to schedule 'delete_replica' jobs to
delete unnecessary replicas from unassigned storages. Replicas pending
deletion must not be elected as the primary as they could accept a write
which gets lost when the scheduled deletion is performed.

This is an edge case and should not happen generally as only assigned
storages can be elected as a primary and only unassigned storages would
have a 'delete_replica' scheduled. However, if an up to date assigned
storage is unassigned, deletion scheduled, the storage is reassigned and
a failover happens promoting the reassigned storage to primary, this could
happen. Fairly unlikely but we have to nonetheless prevent it.