Removal of the config.Config from production code
After a couple rounds of refactorings we can completely
remove config.Config global variable from the production
code. Now it should be passed as a parameter or injected
as a dependency to the places where it is required.
Now we mark it as deprecated.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699