Makefile: revert fd4d7395b and don't pass NO_REGEX=YesPlease
Revert fd4d7395b (makegen: Fix building Git on musl-based platforms,
2020-05-14) and stop passing NO_REGEX=YesPlease.

This means we build with an old version of awk's regex engine, so at
the very least we shouldn't do so outside of musl libc.

But I understand from @pks-t that the musl libc experiment is
something he abandoned, so we can just drop this for now.

As can be seen from the reverted commit the comment I'm removing only
applied to that line, it just became ambiguous in a refactoring that
happened in eace779a9 (Makefile: Get rid of templating, 2020-06-18).