add an index for replication job source replicas
The reconciler checks whether a replica is acting as a source of a
replication job prior to avoid breaking existing jobs when scheduling
'delete_replica' jobs. This commit adds an index for this check to ensure
it doesn't slow down linearly with then number of jobs in the replication
queue.