Fix handling of libgit2 build updates
If the libgit2 version is updated from an existing directory in
`_build/deps/libgit2`, previously we would fetch the branch but the
current version would remain intact. We now call a `git switch` to
switch to the proper branch.

This came out of
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/3181. To test,
switch between `master` and that branch and run `make test`.