testhelper: Refactor protocol helper
The git protocol v2 helper sets up a wrapper which writes environment
data into a temporary file, which is then read via `GetGitEnvData()`.
This interface is not obvious at all, and it depends on having access to
the test directory path. Given that we're about to move the protocol
helper into the `gittest` package, we will not be able to access it
anymore though.

Fix this issue by refactoring the helper function: instead of writing
the data into the test directory directly, we instead create a temporary
directory and put all data in there. The resulting environment file can
then be read via a returned function which replaces `GetGitEnvData()`.