git: Make tests independent of repo teshelpers
In some places, git tests depend on the ability to use the testhelper's
`NewTestRepo()` and related functions. Because of this, we are not able
to use the git package in the testhelper though because it would create
an import cycle. To fix this cycle move all git-related test helpers
into the `gittest` package, but it doe not fix the problem of using
`NewTestRepo()` et al.

Fix this issue by converting the git package to manually create git
repositories and their protobuf representations.