fix racy test TestCoordinator_grpcErrorHandling
TestCoordinator_grpcErrorHandling has a race where in the
"primary error gets forwarded" error case. The mock server does acks
a call as done before setting a flag indicating the node was called.
This causes a race where the primary might return an error before the
secondary calls have been recorded as having arrived. This causes the
test to then flake as the primary error is returned and the assertions
are done before the secondary calls have been recorded.

This commit fixes the issue by recording the call as having arrived
prior to marking decremeting the counter of the WorkGroup.