Makefile: Allow override of default test flags
While we already have an overridable `TEST_OPTIONS` variable in our
Makefile, it doesn't allow a developer to configure test options which
we always pass to tests. Because of this it is neither possible to
override verbosity of test execution as we always pass `-v`, nor is it
possible to override force-execution of tests as we always pass
`-count=1`.

Fix both issues by moving default test flags into our `TEST_OPTIONS`,
allowing users to override them either by passing flags on the command
line or by writing it into the developer's `config.mak` file.