Makefile: Fix git/libgit2 rebuilds causing unshallowing fetch
When either the git or libgit2 version has changed and we already have
cloned the repository, then we stopped discarding the repo in favor of
just doing a git-fetch(1). The fetch invocation doesn't specify a depth
though, which causes us to do an unshallowing fetch which faults in all
objects.

Fix this issue by specifying `--depth 1` for the fetches.