Makefile: add back remote.origin.url for git & libgit2 repos
In c75c02f52 (Makefile: Do not re-clone repos when rebuilding
git/libgit2, 2021-02-17) when this was moved from clone to init/fetch
we lost the remote configuration. It's handy to have when manually
inspecting the repo, so let's add it back.

I'm using "git config" here instead of "git remote add" because as we
potentially add new remote.origin.* config I don't want to deal with
"remote add" punting because we've already configured the remote.