Makefile: supply --no-tags when fetching git & libgit2
As can be seen on the wire with GIT_TRACE_PACKET=1 this saves the
server from spewing a large advertisement of refs/tags/* objects at
us, which we then end up doing nothing with. So we'll fetch the same
data, but this is closer to what we actually want.

That we do this at all is probably a bug in git.git, why would we emit
"ref-prefix refs/tags/" under --depth 1 when asking for an explicit
refs/*/NAME refname? Hrm...

Note that --no-tags doesn't mean "I don't want a tag", but "don't do
tag following".