add replication job target indexes for reconciler
The reconciler checks prior to scheduling update jobs that a target node
is not targeted by by `update` type jobs. It also checks the candidate
source nodes are not targeted by deletions. Prior to scheduling `delete_replica`
jobs, it also verifies the replica is not targeted by any other job.

This commit adds an index that covers all of those checks to improve the query
efficiency.

There's already an existing index for (virtual_storage, target_storage_node) called
virtual_target_on_replication_queue_idx. This index does not contain relative paths,
which forces the query to hit the table anyway. The new index contains all of the
necessary data in the index so we can check only the index.

Parts of the query are aligned with the style in other checks. IN check is changed to
NOT IN to align the check with the index.