Removal of the config.Config from hook package
As we are ready to get rid of the global config.Config
variable this change does exact it for the files in the
hook package. It also includes some small refactorings
like adding a setup function and enabling parallel
execution for some tests.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699