Makefile: set --initial-branch=master on "git init"
Quiet the verbose warning about setting init.defaultBranch in recent
versions of git (due to the master->main transition).

This doesn't matter at all for this use-case of "init", so let's just
explicitly set it to the current default to appease it under V=1, now
that we're not running git with --quiet anymore in that mode.