coordinator: Improve log message when finalizing mutators
We're currently unconditionally logging a "queueing replication jobs"
message when the request finalizer for repository-scoped mutators runs.
This log message is currently misleading though: if there's no outdated
secondaries, no replication jobs will be created.

Reword the log message to "finalizing mutator". In order to be able to
tell whether replication jobs have been created, we now add the primary
as well as the updated and outdated secondaries to the message's fields.