Remove the on-by-default ReferenceTransactions feature flag
The ReferenceTransactions flag has been on by default since
2074eaf50 (Revert "featureflag: Remove reference transaction feature
flag", 2020-12-07), and the non-transaction code for the Ruby codepath
got removed before that in d3ddbdbad (hooks: Remove the Ruby
reference-transaction hook feature flag, 2020-12-04).

It looks like we can remove it now, and since most of the if/else
around keeping this flag was in tests I've worked on actively I was
interested in removing it to speed them up, and make further changes
on them easier.