Introduction of the Deferrer
Deferrer allows to setup list of the function to be
called with Call method or a new instance could be created
with all set functions, but the old will be left an empty.
It helps a lot when you need to setup something and collect
a set of cleanups, but not call them in a setup function and
return back to the caller. And at the same time they should
be called if some setup step failed.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699