align StorageProvider's interface with RepositoryStore's
StorageProvider is the interface of the the cache that is caching
the consitent storages. The interface deviates from RepositoryStore's
GetConsistentStorages. This necessiates an additional type
DirectStorageProvider for just adapting GetConsitentStorages to the
changed interface. More so, the differing interface is causing additional
loopin later in nodes.Manager's GetSyncedNode as we have to loop over the
returned slice of storages instead of just looking up a key from the map
returned by GetConsistentStorages.

There seems to be no reason as to why the alternative interface would be
preferred, so this commit aligns the StorageProvider's interface with the
one provided by RepositoryStore.