cache: Fix racy usage of mocked counters
Starting with the introduction of the new Gitaly config builder,
`TestDiskCacheObjectWalker()` has started to become racy. This is
because when creating a new Gitaly configuration via the config builder,
we'll call the `Validate()` function. This function not noly validates
the configuration, but also invokes a batch of hooks. Where one hook is
the cache walker, which gets released in order to clean up the cache.
The cache walker races with our use of the mocked cache mock counters:
it accesses the counter while we change its pointer to point to a
different mock pointer in order to reset it.

Fix the race by not changing the mock counter's address, and instead
calling a new `Reset()` function which uses proper locking.