praefect: Add `forcePrimary` parameter to `RouteRepositoryAccessor`
Implementations of the Router interface are all forced to check
themselves whether they must force-route accessors to the primary node
because the RPC's metadata tells them to. While annoying, this hasn't
been much of a problem until now given there's only two implementations.
But we need to extend the conditions to also take into account what kind
of RPC we are currently routing. Currently, the Routers are designed
RPC-agnostic though except for an RPC's type, and this should stay this
way. Instead, it's the coordinator which should inspect the RPC and tell
the router to force-route to primaries.

As a preparation for this change, this commit adds a new `forcePrimary`
flag to the `Router` interface. If set, it tells the router that it must
always route to the primary node. No change in behaviour is expected
from this change as we always pass `false` for now.