Makefile: Add benchmarking target
We do not yet have any well-defined benchmarking best practices in the
Gitaly team. This commit thus introduces a new `bench` Makefile target
which clones a reference repository (gitlab-org/gitlab to benchmark
against). This is not an ideal benchmarking target given that the
repository changes over time. But it's a first iteration towards a
unified benchmarking strategy.