blob: Cancel execution of git-rev-list(1) when hitting LFS pointer limit
When searching for LFS pointers, we build a pipeline which writes from
git-rev-list(1) into git-cat-file(1). In order to verify that the
pipeline has finished successfully without any data being dropped
because of an error, we need to check both head and tail of the pipeline
for any errors. To do so, we thus call `Wait()` for both commands: once
when we've got an EOF in git-cat-file(1), and once when git-rev-list(1)
is supposed to exit.

There is an exception to this though: when the loop reading from
git-cat-file(1) hits the limit of requested LFS pointers, it will return
early without waiting for it to exit. This makes perfect sense: we do
not care for an error, we've already got everything the user wants. But
we still wait for the upstream git-rev-list(1) to terminate its graph
walk. This is obviously needless busywork: we do not care for its error
code either, we still got everything the user asked for.

Fix the issue by just not waiting for git-rev-list(1) to exit in case
we've got all LFS pointers the user requested. This severely improves
performance when a limit is in place when walking repos with a big
revision graph:

    # Before
    BenchmarkFindLFSPointers/limitless-16         	       1	20224658688 ns/op
    BenchmarkFindLFSPointers/limit-16             	       1	14362972587 ns/op
    BenchmarkReadLFSPointers/limitless-16         	       1	18990121780 ns/op
    BenchmarkReadLFSPointers/limit-16             	1000000000	         0.113 ns/op

    # After
    BenchmarkFindLFSPointers/limitless-16         	       1	20154639708 ns/op
    BenchmarkFindLFSPointers/limit-16             	1000000000	         0.180 ns/op
    BenchmarkReadLFSPointers/limitless-16         	       1	18473904116 ns/op
    BenchmarkReadLFSPointers/limit-16             	1000000000	         0.108 ns/op

That's definitely a huge speedup. I somehow doubt that this is accurate,
but it's reproducible.