blob: Check exit status of git-cat-file(1) when reading LFS pointers
When reading LFS pointers via git-cat-file(1), we're currently looping
the command's stdout until we get either an unexpected error or an EOF.
In case we've got an EOF, we cannot tell though whether it's because the
command exited successfully or not based on the EOF alone.

Fix the issue by properly waiting for the command to terminate after the
loop.