blob: Drop rogue optimization when reading LFS pointers
Git does not currently provide a way to find all blobs referenced by a
range of revisions. While there is the `--filter=blob:limit=` filter to
git-rev-list(1) to limit the list of candidate LFS pointers, the biggest
downside of it is that it'll still print all other object types.

To work around this limitation, we have grown an optimization in our
Ruby code which uses the `--object-name` switch as a heuristic to deduce
whether a given object can be a blob or not: blobs always have an object
name, which is simply their path as seen first when doing the graph
walk. But this optimization isn't ideal in any way: both tags and trees
also have names, namely their tag name and the tree paths for everything
but the root tree. So effectively, all we gain from this optimization is
that we can filter out commits.

But as it turns out, this optimization is even a rogue one. Because we
now have to filter all object IDs by whether they have a path or not, we
have grown an extremely inefficient access pattern: we take candidate
objects returned git-rev-list(1) and request them one by one from the
running git-cat-file(1) instance. On each iteration, we thus have a
roundtrip of writing to git-cat-file(1) and then waiting for it to
return the object info.

Fix this by directly wiring up git-rev-list(1) and git-cat-file(1). This
requires us to get rid of the `--object-name` optimization, but as can
be seen by the following benchmarks that's outweighed by the improved
access pattern:

    # Before
    BenchmarkFindLFSPointers/limitless-16         	       1	42812541563 ns/op
    BenchmarkFindLFSPointers/limit-16             	       1	14662540521 ns/op
    BenchmarkReadLFSPointers/limitless-16         	       1	47841769495 ns/op
    BenchmarkReadLFSPointers/limit-16             	1000000000	         0.362 ns/op

    # After
    BenchmarkFindLFSPointers/limitless-16         	       1	20224658688 ns/op
    BenchmarkFindLFSPointers/limit-16             	       1	14362972587 ns/op
    BenchmarkReadLFSPointers/limitless-16         	       1	18990121780 ns/op
    BenchmarkReadLFSPointers/limit-16             	1000000000	         0.113 ns/op

So we've roughly halfed/tripled execution times in cases where we do a
complete graph walk and thus request a whole bunch of objects.